# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Tangula API接口文档\n唐古拉山脉，长江正源，滋润万物。\nTangula 寓意数据的源头，为研发团队提供即取即用的数据服务，帮助团队高效产出。",
        "title": "Tangula",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "zhang.jiejun@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "paths": {
...
        "/script/browse": {
            "get": {
                "description": "【登录权限】浏览脚本列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "脚本管理"
                ],
                "summary": "浏览脚本列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "分页索引",
                        "name": "index",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 15,
                        "description": "分页数量",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "名称搜索过滤参数",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/script.Script"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/script/upload": {
            "post": {
                "description": "【登录权限】上传脚本。",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "脚本管理"
                ],
                "summary": "上传脚本",
                "parameters": [
                    {
                        "type": "string",
                        "description": "目标脚本文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "脚本描述",
                        "name": "desc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "$ref": "#/definitions/script.Script"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/script/{id}/content": {
            "get": {
                "description": "【登录权限】查看脚本内容。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "脚本管理"
                ],
                "summary": "查看脚本内容",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "指定脚本ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Result"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/script/{id}/delete": {
            "delete": {
                "description": "删除脚本。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "脚本管理"
                ],
                "summary": "删除脚本",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "指定脚本ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Result"
                        }
                    }
                }
            }
        },
        "/script/{id}/download": {
            "get": {
                "description": "下载脚本。",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "脚本管理"
                ],
                "summary": "下载脚本",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "指定脚本ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "response": {
                    "type": "object"
                }
            }
        },
...
        "script.Script": {
            "type": "object",
            "properties": {
                "create_user": {
                    "type": "string"
                },
                "created_time": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
    }
}